Amazon Audit Grace Period
IN some cases the queue consumer is not idempotent. One of these
cases is where a properly collected message is present in the
queue and an audit collected message is then added. These will
have different timestamps and thus will be treated as different
transactions.

We remove this possibility by adding a grace period where the audit
script will not check for messages before a certain time.

Change-Id: If136e2b3dce49f6635eb968b429d07a8981ce229