Logging every Transaction During Audit
I have a feeling that Amazon is lying to me about what transactions
it has when I call getAccountActivity(). This will log all transactions
that we download so I can verify that particular point.

It also forces input timestamps into UTC which might solve a bug
where it looks like Amazon is treating some times as PDT but others as
UTC.

Change-Id: Id4befe8921dc397f2968fe40a436bd545070eb32