Rough YAML linter for *.yaml.example files
Iterate the repository for files matching '.yaml' or '.yaml.example' and
attempt to safe_load() them.

YAML files containing multiple documents are not supported, only
the first document will be validated. That is because safe_load_all()
does not raise any error for some reason.

Add test-requirements.txt to ship nose and PyYAML.
Introduce tox env 'yamllint' and add it to the default list of env to
run (envlist).

Usage is either:

 tox
 tox -eyamllint

Bug: T114470
Change-Id: I076235213a81fd6431919f28fcead6c72f5cabb2