Alter definition of silverpop_countrylangs to match actual
I have been hitting
_mysql_exceptions.OperationalError: (1267, Illegal mix of collations (utf8_unicode_ci,IMPLICIT) and (utf8_general_ci,IMPLICIT) for operation '=')

so I figured the table definition should be updated to match what is actually in use. It turned out to be a little
surprising - ie. varbinary & charset=binary, but I guess that is what the actual DB is using so it should be what
is in our code

Change-Id: Ic79e745028f32542de900f4819529a86124a648b