Fix unsafe reference to config hash
Now that I understand Python imports a bit better...  The bare "config"
variable is unsafe to import, because we're depending on it having been
initialized before the module using it is itself imported, which is fragile and
hard to predict.

This patch uses the get_config function at the point of usage, which should be
more dependable.

The config system needs more cleanup, e.g. now that we use the safer get_config,
we can push load_config into entry point's methods rather than including in the
top-level import sequence.

Change-Id: I7377058ee324d4dc5fda20c185739ce50a59a714