Fix miscalculated annual fields, inclusion of endowment
On digging into the endowment exclusion I found a serious error in
the code we deployed last week. The inclusion of contributions in
the left join is leading to inaccurate amounts. Tests show that we
can speed this up as well by keeping contributions out of the join.

I've simply added financial type in where we are still relying on
the contribution table (because currently we are not collected
date of largest contribution)

Bug T228349
Change-Id: I07fd9f714e04f9f85212e651adfeb53a1e419218