Add new Silverpop fields to export.
This is expected to be temporary but let's see how long it takes in the  first instance.

Complexities are
1) co-ordination with online - if  we add fields  at the end  does that mean we can add them with  looser co-ordination?
What about if they start with Z
2) Mappings -  should we map the fields to their option values - I did this for Gender
3) Deduping - we don't quite know what happened with ts data where we sent them lists. Spot checks show in most
cases we have (the same) prospect fields for both but in some cases not. It seems likely  that DS will do some targetted
deduping on our more interesting targetsmart results so at this stage we are just grabbing the data
as it relates to the exported contact id & we can monitor how long that adds to the import (do we leave it
in the import or remove again) and also try to ascertain how much data might be missed

Bug: T231538
Change-Id: I761294577a43ec0237d259e3fae4559fe8b787fe