Add to, rather than rebuild the suppression list.
Building the master suppression list takes about an hour. In testing on staging these queries took less than a minute
(although more at peak times I presume).

I think it really is this simple... if we don't drop the table the we just need to add any possibly-changed
values to it and then let our normal delete do a quick cull to re-remove any opted in.

 I wrote my analysis in the code comments - please attend to any caffeine needs before proceeding
to review.

Bug: T253152

Change-Id: I2eddeb0a802cd9f0f5b0d785012e0322605a9892