Silverpop refactor - reduce queries to build country & language
This takes about the same amount of time before & after. The reason for doing it is that there is a goal of
getting rid of updates to tables, so that the table is always in a consistent state. Otherwise,

when we switch to incremental updates we have several queries acting on the same piece of data in the same row

Change-Id: I5b494c119e3ff555b23bd69094ee399b1562d695