Add endowment_latest and endowment_highest fields
This change adds all the MAX fields to the stats table. It then uses this table to create
a table of endowment_latest stats and endowment_highest stats.

The change to the stats table adds 10 minutes onto
this table's creation. However, in the process it generates the lastest donation date
for all donations. I intend to leverage this field in later updates
to treat the email associated with the latest donation (rather than the
highest email id) as the master.

Note I don't think
1-to-1 left joins in final view as terribly problematic and I think they are
faster than moving the data around. In this case it's a small-ish table.

This commit replaces some placeholders with actual data

Bug: T252245

Change-Id: I1719a5cbf16f3a2817b5affa1624a1b6cf455056