Silverpop refactor - remove last UPDATE from the silverpop script
This removes an update query & it seems that it's quicker without it too (although I do get slightly
variable results depending on the live silverpop job running. This kinda allows me to move the focus
from 'getting rid of updates' to 'incrementalism'. Dropping the email->key->update
seems to take 6 minutes off the query time and the dedupe join should make it obsolete

Bug: T253152

Change-Id: Ica2bef09165339b867d7240b99a1ffb1010ae729