Use summary table to determine opted in
We are seeing a situation where emails with multiple contact records have been removed
from the the suppression list if any one of the contacts associated with the list is not
opted out and is not-not opted in.

The delete is removing based on ANY match rather than ALL matching.

It regressed because we used to purge some addresses from the table but not leave them in.

This switches to using the summary table - with a calclated overall value.

Note that all contacts altered in the last 7 days will be re-added to the table. Anyone older would already be
on the suppression list I think but I upped it to 14 days temporarily to make sure none got missed.

I think the deploy that changed it is much more recent...
Bug T253152

Change-Id: I2d3763726a508b6b9f2e0dd42f81ac19c8873540