Add all funds last donation date to export_staging table
I think we need to determine master email id off the latest donation date & only the highest email id if the former
is not applicable. I've done quite a bit of (tedious) query testing and with the added index we can get good performance
on calculating the email which maps to this latest value. However, we need to add it and populate it before
any incrementalism on the staging table

ie the query is

-- Query OK, 23208067 rows affected (4 min 25.47 sec)

CREATE TABLE silverpop_email_map
  SELECT
    s.email,
    COALESCE(MAX(stat.id), MAX(s.id)) as master_email_id,
    COALESCE(MAX(stat.address_id), MAX(s.address_id)) as address_id,
    COALESCE(MAX(stat.preferred_language), MAX(s.preferred_language)) as preferred_language,
    MAX(s.opted_out) as opted_out,
    # 0 if they have ever actually opted out, else 1
    # we use this for filtering so don't need to preserve the nuance.
    # This should be revisited per https://phabricator.wikimedia.org/T256522
    MIN(IF (s.opted_in = 0, 0, 1)) as opted_in
  FROM silverpop_export_staging s
     LEFT JOIN silverpop_export_staging stat
       ON stat.email = s.email
       AND stat.all_funds_latest_donation_date = s.all_funds_latest_donation_date
  GROUP BY email
;
 It is 7 minutes faster due to the index

Bug: T253152
Change-Id: Ie549a4b0a58e313fb05a16fd552555b5c7596697