Remove temporary extra-primary delete
I'm a bit nervous about these lines & I think they were a bit of a stop-gap. The goal was really about
protecting against a bit of a split-second scenario (which I think we avoid through timing the jobs
and quicker queries now). On the other hand this opens us up to people winding up on the suppression
list when they shouldn't I think. Because we build the table about 30 minutes before this it could be
that a contact whose email id is deleted and replaced in those 30 minutes has their
deleted email removed from silverpop_export_staging but their new email not added
until the following day. I think this could see it added to the suppression list.

On balance I'd rather risk hard-fails returning (with failmail) than too much suppression

Change-Id: Id796f16025cfedb253936640f6925a90a0c8d7f9