Add annual total fields to everywhere except the final view
These fields are not output (and so we don't need to co-ordinate with Katie)
until we add them to the final view - this adds them to all places short of that.

Note that Katie has requested a new naming convention in the final view
- see https://phabricator.wikimedia.org/T283798#7272217 (but that is after
we agreed changing foundation to AF) so the 3 added fields are there.

Also note I've done 2 years into the future as I plan to add those fields to the
live db as well so we get a reprieve before adding more

Bug: T280595
Change-Id: I57a689006104419bd656aa475d07c0bbccef259c