Update Silverpop config files to reflect our docker credentials.
Document current state - note that we should improve on this set up - but I think this does
a reasonable job of providing enough information to get the code to actually run and
we should probably focus on iterating here rather than getting it right (I say this
because I first wrote a lot of this in March 2021 & I have been treating my
personal WIP patch as 'the best source of documenation' for over
2 years now.)

Also note that I removed one requirement that would not install for me.

This requirement was added in 2016 in patch 75b422a9d9fc882f30f808789873b9905be2ceb2
I took a look at the patch and it was moved from a silverpop-specific requirements
file to a general requirements file at that point. Ergo we can have some confidence
it won't hurt anything if it doesn't hurt the silverpop export.

Further the setup really only happens on jenkins and in ci. Any python
package installations on prod or staging are managed by fr-ops. So, if it works
locally per the Readme instructions and on jenkins then we can be confident

Bug: T341017

Change-Id: I5eb8105ab4f281bae04737e46051a0457f80b759