Remove receive_date FORCE index
This query sped up significantly without the FORCE

The explain shows that with the FORCE is was using civicrm_contribution
as the base table & it switches to using silverpop_update_world
without it. I feel like maybe this force was in place before we
switched to using the update_world constraint & maybe other index hints
got removed but this was missed? It seems odd for it to be on this
table & not the others.

Regardless - changing data can change how various queries perform
and it might have made sense once..
Bug: T347997
Change-Id: I6e8c2011ab16f00640fe20c6e7220d8c2ad41fa9

Change-Id: I32ce9747f511b1453b7c0e4d1b8660dacae05e24