Update silverpop with removal of fields to be removed
Per https://phabricator.wikimedia.org/T347724#9255989
this removes the calendar year fields from the acoustic export but
in the final view it zeros them out rather than removes them.

The schema changes won't actually need to be made pro-actively
(ie if the fields are not in WMF donor but as still in the silverpop
schema for a while that's fine cos the script won't be
re-populating them on the silverpop side either way.
However, I did add in the new segment & status fields on
the silverpop schema so when we do touch it they
will be part of the change

Bug: T347724
Change-Id: I4f06c753a5c2033da535f3fcb9891525c3154595