Replace EXECUTOR_NUMBER env var by CI
The CI test suite uses docker-registry.wikimedia.org/releng/tox-mysqld
which spins a MariaDB server using Debian default settings.

To configure the database setting under CI, we checked whether
EXECUTOR_NUMBER environment variable is set. That is specific to
Jenkins. We are now using `CI` which is agnostic to the system running
CI.

Bug: T399283
Change-Id: I237318141eff0a5a5ea14cf03282b61ff50a144f