Add option of full rebuild of suppresion_list.
No change to any current behaviour, but this gives us the option of rebuilding the suppression list from zero. This will make it feasible to check if opted out contacts in Acoustic (which we uploaded) should actually be opted on a large scale by looking for them in the silverpop_excluded.
Added test shows that this behaves the same as the current incremental approach, at least as far as the cases tested.
We can use this to re opt-in contacts that we removed, but not to contacts opted out directly in Acoustic as some of this data is currently missing in CiviCRM (T411896) and I don't think present in CiviCRM as we go back further in time, so that process will require additional work.

Bug: T409958
Change-Id: Id7c1a5da3d9fff785da6a1699eb7714b972c3fd0